checkpoint de agregar metodos de conte por harware

+++ src/Contexts/employee/Unidad/application/UnidadRemover.ts
@@ -1,15 +1,19 @@
 import { UnidadId } from '../domain/valueObject/UnidadId'
-import { type Primitives } from '../../../Shared/domain/value-object/Primitives'
-import { type UnidadRepository } from '../domain/repository/UnidadRepository'
+import type { Primitives } from '../../../Shared/domain/value-object/Primitives'
+import type { UnidadRepository } from '../domain/repository/UnidadRepository'
+import type { EventBus } from '../../../Shared/domain/event/EventBus'
+import { UnidadDoesNotExistError } from '../domain/errors/UnidadDoesNotExistError'
+import { Unidad } from '../domain/entity/Unidad'
 
 /**
  * @description Use case for removing an Unidad entity.
  */
 export class UnidadRemover {
 	private readonly unidadRepository: UnidadRepository
-
-	constructor({ unidadRepository }: { unidadRepository: UnidadRepository }) {
+	private readonly eventBus: EventBus
+	constructor({ unidadRepository, eventBus }: { unidadRepository: UnidadRepository; eventBus: EventBus }) {
 		this.unidadRepository = unidadRepository
+		this.eventBus = eventBus
 	}
 
 	/**
@@ -19,6 +23,14 @@ export class UnidadRemover {
 	 */
 	async run({ id }: { id: Primitives<UnidadId> }): Promise<void> {
 		const unidadId = new UnidadId(id).value
+		const unidad = await this.unidadRepository.findById(unidadId)
+
+		if (!unidad) {
+			throw new UnidadDoesNotExistError(unidadId)
+		}
+		const unidadEntity = Unidad.fromPrimitives(unidad)
+		unidadEntity.delete()
 		await this.unidadRepository.remove(unidadId)
+		await this.eventBus.publish(unidadEntity.pullDomainEvents())
 	}
 }
